# docs/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    The PaymentFlow payments API. This document describes the public `/v1` surface — the endpoints PaymentFlow versions, deprecates on a published timeline, and supports for integrators. Endpoints not described here are internal and carry no compatibility promise.

    Amounts are integers in the currency's minor unit (`amountMinor`): `1000` in `USD` is $10.00. There are no floating-point amounts anywhere in this API.

    Every mutating request accepts an `Idempotency-Key` header and every list is cursor-paginated; both are described on the operations that use them.
  license:
    name: Proprietary
  title: PaymentFlow API
  version: 2026-08-01
servers:
- url: https://api.paymentflow.dev
  description: PaymentFlow API
security:
- SecretKey: []
paths:
...
  /v1/test/simulations:
    post:
      description: |-
        Installs an override that makes the sandbox behave a given way for your next authorizations, regardless of which test card they use.

        Test cards cover the outcomes you can choose per payment; an override covers the ones you cannot — a processor outage, a sudden run of declines, latency that trips your timeouts — without needing a token for every combination. Bound it with `remainingCount` or `durationSeconds` so it stops on its own.

        One override is active per merchant and mode at a time; creating another replaces it.
      operationId: createSimulationOverride
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSimulationOverrideRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationOverrideResponse"
          description: The override is active.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "The override is not internally consistent — a `FORCE_DECLINE` with no `declineCode`, or neither `remainingCount` nor `durationSeconds` to end it."
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: Force a sandbox behaviour
      tags:
      - Simulations
  /v1/test/simulations/active:
    delete:
      description: "Removes the override in force for your merchant and mode, returning the sandbox to its normal behaviour. Succeeds whether or not one was active — the point is the resulting state, so this is safe to call unconditionally at the start of a test."
      operationId: revokeActiveSimulationOverride
      responses:
        "204":
          description: No override is in force. Returned whether or not one was.
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: Revoke the active override
      tags:
      - Simulations
    get:
      description: "Returns the override currently in force for your merchant and mode, including how much of it is left. The first thing to check when the sandbox is behaving in a way you did not ask for — an override with `remainingCount` left over from an earlier test is the usual cause."
      operationId: getActiveSimulationOverride
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationOverrideResponse"
          description: The active override.
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: No override is active — the sandbox is behaving normally.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: Retrieve the active override
      tags:
      - Simulations
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: "A stable, machine-readable identifier for this specific failure, such as `PAYMENT_NOT_CAPTURABLE`. The set of codes grows by policy, which is why `type` and not this is what a client should switch on. Every code is catalogued in docs/ERRORS.md."
        correlationId:
          type: string
          description: "Identifies the whole distributed trace this call belongs to, which may span several services. Broader than `requestId`."
        docUrl:
          type: string
          description: Where to read about this specific code.
        errors:
          type: array
          description: "Field-level validation failures, when the request failed validation in more than one place."
          items:
            $ref: "#/components/schemas/ApiFieldError"
        message:
          type: string
          description: "A human-readable explanation. Safe to log; not intended to be shown to your customers, and never programmatically parsed."
        param:
          type: string
          description: "The single offending parameter, when there is exactly one. Multi-field validation failures use `errors` instead."
        path:
          type: string
          description: The request path that produced the error.
        requestId:
          type: string
          description: Identifies this one HTTP call. **Quote this in a support request** — it is on every log line and every row of your request log.
        status:
          type: integer
          format: int32
          description: "The HTTP status code, repeated in the body so a logged response is self-contained."
        timestamp:
          type: string
          format: date-time
          description: "When the error occurred, as RFC 3339."
        type:
          type: string
          description: "The error's classification, and **the field to branch on**. This is a small closed set an SDK maps to exception classes: `authentication_error`, `permission_error`, `invalid_request_error`, `idempotency_error`, `rate_limit_error`, `api_error`."
    ApiFieldError:
      type: object
      properties:
        field:
          type: string
          description: The request field that failed validation.
        message:
          type: string
          description: Why it failed. The rejected value is deliberately never echoed back.
...
    CreateSimulationOverrideRequest:
      type: object
      properties:
        declineCode:
          type: string
          description: "The decline code to return. Required for `FORCE_DECLINE`, ignored otherwise."
          example: insufficient_funds
          maxLength: 48
          minLength: 0
        durationSeconds:
          type: integer
          format: int32
          description: "How long the override lasts, in seconds. The alternative to `remainingCount`; supply one of the two."
          example: 300
        errorCode:
          type: string
          description: "The error code to fail with. Required for `FORCE_ERROR`, ignored otherwise."
          example: processor_unavailable
          maxLength: 48
          minLength: 0
        latencyMs:
          type: integer
          format: int32
          description: "How long to stall, in milliseconds. Required for the latency scenarios, ignored otherwise."
          example: 3000
        remainingCount:
          type: integer
          format: int32
          description: "How many authorizations the override applies to before it expires. **Supply this or `durationSeconds`** — an override with neither would never stop, and a sandbox you cannot get back out of is worse than one you cannot get into."
          example: 3
        scenario:
          type: string
          description: "Which behaviour to force. The scenario decides which of the fields below are required: `FORCE_DECLINE` needs `declineCode`, `FORCE_ERROR` needs `errorCode`, and the latency scenarios need `latencyMs`."
          enum:
          - FORCE_DECLINE
          - FORCE_ERROR
          - INJECT_LATENCY
          - FORCE_TIMEOUT
          - FORCE_RATE_LIMIT
          - DELAY_SETTLEMENT
          - DUPLICATE_WEBHOOKS
          - WEBHOOK_FAILURE
          example: FORCE_DECLINE
      required:
      - scenario
...
    SimulationOverrideResponse:
      type: object
      properties:
        declineCode:
          type:
          - string
          - "null"
          description: "The decline code being returned, for a decline scenario. **Null** otherwise."
          example: insufficient_funds
        enactedFrom:
          type:
          - string
          - "null"
          description: "Which part of the platform acts on this scenario. **`null` means the decision engine enforces it now**; a value names the release that will. The webhook scenarios are stored and validated but not yet acted on, and saying so here is more honest than accepting an override that silently does nothing."
          example: M18
        errorCode:
          type:
          - string
          - "null"
          description: "The error code being returned, for an error scenario. **Null** otherwise."
          example: processor_unavailable
        expiresAt:
          type:
          - string
          - "null"
          format: date-time
          description: When the override stops applying. **Null** for a count-bounded override.
        id:
          type: string
          format: uuid
          description: Unique identifier for this override.
        latencyMs:
          type:
          - integer
          - "null"
          format: int32
          description: "The latency being injected, in milliseconds. **Null** for a scenario that injects none."
          example: 3000
        remainingCount:
          type:
          - integer
          - "null"
          format: int32
          description: How many authorizations the override still applies to. Counts down as it is used; **null** for a time-bounded override.
          example: 2
        scenario:
          type: string
          description: The behaviour being forced.
          example: FORCE_DECLINE
...
  securitySchemes:
    SecretKey:
      bearerFormat: sk
      description: |-
        Your secret API key, sent as `Authorization: Bearer sk_test_...`.

        The key determines both *whose* data you see and *which mode* you see it in: an `sk_test_` key reads and writes test data only, an `sk_live_` key live data only. No header, parameter, or body field can change either — to work in the other mode, use the other key.
      scheme: bearer
      type: http